feat: add working modal with extra info
